Fix MIDI pause/stop functionality (fluidsynth player only)
Now stop will actually reset to the start, and pause will properly resume on pressing play again.

Looks like fluidsynth can seek now, however it goes by MIDI ticks which might be tricky as we go by milliseconds. If a MIDI has tempo changes we can't really sync the seek bar and the song properly. Might look into it later because it'd be nice to be able to seek when playing MIDIs.