Program database initial commit
I like the idea from the other 3.3.0 experimental branch/PR #1542 but I'd like to keep it as simple as possible initially, so let's get a basic implementation going in 3.2.X to start with. It'll basically be the same as the database_330 branch just without the more complicated 'archive library' stuff.

What will be in:
- Window info and layout
- UI state variables
- Previously opened archive files
- Per-archive UI state (column sizes, run config, etc.)

What won't be in:
- The 'Archive Library' UI
- Full archive entry info

Undecided:
- Archive bookmarks
- Archive map info