Game configuration JSON part 1 (WIP) - base game configurations
Converted all the base game configurations to a JSON format. Where previously we used #include, now we use an 'import' section which works much the same way.

Port configurations will be a bit more difficult as they have a lot of conditional parts, previously handled via #define/#ifdef/#endif etc. Since we can't do that now (at least, I'd rather not need to do extra processing of the JSON files), we'll need a different way to handle it. Ideally I'd like to keep all the conditions in the base port configuration files, and not littered throughout the included files