Add button to switch entry list view type on the fly
Now we no longer need to change it in settings then reopen the archive. The setting is now saved to the UI state table instead (will make it saved per-archive later). The old cvar has been removed and is migrated to the UI state on first run

Not 100% happy with the button icons but they will do for now