Initial work on exception handling via cpptrace
With cpptrace we can get a stack trace for an exception, as long as it was generated between CPPTRACE_TRY/CPPTRACE_CATCH and not just a regular try/catch. Unfortunately we can't just modify wx to use these internally so for now I've just used them in some 'common' places like doAction. To cover everything we'd need to use them in every wx event handler which is unrealistic (and probably overkill)

Also added a custom exception dialog and all unhandled exceptions will pop it up (instead of the default wx one)