Use libpng directly for png support instead of wx (#1916)
* Use libpng directly for png support instead of wx

When writing wxImage to png, it will always condense the palette if one exists, which breaks a bunch of things eg. colour translation.

This reworks SIFPng to use libpng directly to load/save png images, instead of wxImage

* Fix non-msvc build error

* Use system libpng when building wx

By default it uses it's own version, but if they mismatch we get a crash on loading pngs with wx